connectivity: Move template utils from internal
Expose package template methods to allow connectivity test hooks
rendering yaml templates.
